// CloneFE/frontend/Frontend/src/components/partial/TrainingProgramManagement/TrainingProgramList/models/TrainingProgram.model.ts
import { Syllabus } from "../../CreateTrainingProgram/models/syllabus.model";

export interface TrainingProgram {
  id: number;
  trainingProgramCode: string;
  name: string;
  createdDate: Date;
  createdBy: string;
  updateDate: Date;
  updateBy: string;
  startTime: number;
  days: number;
  hours: number;
  status: string;
  syllabi: Syllabus[];
}